moved sunburst polyfill to separate file
