+++ src/components/visualizations/sunburst-vega/sunburst-vega.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable no-magic-numbers */
 import React from 'react';
 import { Vega } from 'react-vega';
+import appendPolyfill from './utils/append-polyfill';
 import sunburstSpec from './utils/sunburst-spec';
-
 import './sunburst-vega.scss';
 import PropTypes from "prop-types"
 
@@ -50,36 +50,12 @@ export default class Sunburst extends React.Component {
   render() {
     const { data, spec } = this.state;
     if (typeof window !== 'undefined') {
-	// polyfill for vega library
-	// Source: https://github.com/jserz/js_piece/blob/master/DOM/ParentNode/append()/append().md
-		(function (arr) {
-		  arr.forEach(function (item) {
-			if (item.hasOwnProperty('append')) {
-			  return;
-			}
-			Object.defineProperty(item, 'append', {
-			  configurable: true,
-			  enumerable: true,
-			  writable: true,
-			  value: function append() {
-				var argArr = Array.prototype.slice.call(arguments),
-				  docFrag = document.createDocumentFragment();
-        
-				argArr.forEach(function (argItem) {
-				  var isNode = argItem instanceof Node;
-				  docFrag.appendChild(isNode ? argItem : document.createTextNode(String(argItem)));
-				});
-        
-				this.appendChild(docFrag);
-			  }
-			});
-		  });
-		})([Element.prototype, Document.prototype, DocumentFragment.prototype]);
-      return (
-        <Vega data={data} spec={spec} signalListeners={this.signalListeners} />
-      )
+		appendPolyfill();
+		return (
+			<Vega data={data} spec={spec} signalListeners={this.signalListeners} />
+		)
     } else {
-      return <div></div>
+		return <div></div>
     }
   }
 }
